fix(vault): secret is temporarily empty after changed vault config
When vault config is changed, in the worker event we first flush the LRU
cache, then start to update the secrets from vault provider. There’s a
period of time in between that the cache is empty.
The `kong.vault.update()` function only lookups cache and will update
the secret to an empty string when cache is empty. This can cause plugins
to throw nil errors. This commit changed the worker event callback to
not flush the LRU cache. The cache will be updated when the secrets are
fetched from vault prodiver.