fix the update logic to change values if no error
The test "resurrects plugin config references when secret is deleted" failed.
It is asserting that when a secret resurrect ttl is passed, it should be
deleted. However with previous update() function logic the value is
only set if it's not nil. This commit changes the logic to set the value
based on if there's an error or not.